refactor: move createSession to session handler module
